# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: pc-agent360 Backend API
  version: 1.0.0
  description: >-
    API para autenticación Firebase, sesiones opacas y administración de acceso.
    Los errores son genéricos y nunca contienen tokens, cookies, correos ni trazas.
servers:
  - url: http://127.0.0.1:3000
    description: Desarrollo local
  - url: https://api.example.com
    description: Referencia de despliegue HTTPS
paths:
...
  /api/v1/admin/users:
    get:
      tags: [Users]
      summary: Lista la whitelist de usuarios autorizados
      description: Requiere el rol ADMIN y el permiso users:read actual.
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Usuarios autorizados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Users]
      summary: Agrega un usuario a la whitelist
      description: Requiere el rol ADMIN, users:write y un CSRF válido.
      operationId: createUser
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Usuario autorizado creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/admin/users/{userId}:
    get:
      tags: [Users]
      summary: Consulta un usuario de la whitelist
      description: Requiere el rol ADMIN y el permiso users:read actual.
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/UserId'
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Usuario autorizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Users]
      summary: Actualiza usuario, rol o estado de la whitelist
      description: Requiere el rol ADMIN, users:write y un CSRF válido.
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/UserId'
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuario actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Users]
      summary: Elimina un usuario de la whitelist
      description: Requiere el rol ADMIN, users:write y un CSRF válido.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/UserId'
      security:
        - SessionCookie: []
      responses:
        '204':
          description: Usuario eliminado y sus sesiones invalidadas.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/admin/roles:
    get:
      tags: [Roles]
      summary: Lista roles y permisos permitidos
      description: Requiere el rol ADMIN y el permiso roles:read actual.
      operationId: listRoles
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Roles disponibles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesEnvelope'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Roles]
      summary: Crea un rol personalizado
      description: Requiere el rol ADMIN, roles:write y un CSRF válido.
      operationId: createRole
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleRequest'
      responses:
        '201':
          description: Rol creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/admin/roles/{roleId}:
    get:
      tags: [Roles]
      summary: Consulta un rol
      description: Requiere el rol ADMIN y el permiso roles:read actual.
      operationId: getRole
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/RoleId'
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Rol encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Roles]
      summary: Actualiza un rol personalizado o los permisos del rol ADMIN
      description: Requiere el rol ADMIN, roles:write y un CSRF válido. En el rol protegido ADMIN solo se puede actualizar la lista de permisos.
      operationId: updateRole
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/RoleId'
      security:
        - SessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleRequest'
      responses:
        '200':
          description: Rol actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Roles]
      summary: Elimina un rol personalizado no asignado
      description: Requiere el rol ADMIN, roles:write y un CSRF válido.
      operationId: deleteRole
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/RoleId'
      security:
        - SessionCookie: []
      responses:
        '204':
          description: Rol eliminado.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    FirebaseIdToken:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
      description: Token Firebase en memoria, usado solo para crear o renovar sesión.
    SessionCookie:
      type: apiKey
      in: cookie
      name: agent360_session
      description: Nombre por defecto; puede cambiar con SESSION_COOKIE_NAME.
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      description: UUID v4 para correlacionar la solicitud. Se genera uno si falta o es inválido.
      schema:
        $ref: '#/components/schemas/CorrelationId'
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Token CSRF entregado por la creación, bootstrap o renovación de sesión.
      schema:
        type: string
        minLength: 43
        maxLength: 43
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV4'
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/UuidV4'
...
  responses:
    BadRequest:
      description: Solicitud con formato, tipo o campo no permitido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthenticated:
      description: Falta, expiró o no es válida la identidad o sesión.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: La sesión no tiene el rol ADMIN, el permiso o el origen requerido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: El estado actual no permite procesar la solicitud.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Se superó el límite temporal de solicitudes.
      headers:
        Retry-After:
          description: Segundos mínimos antes de reintentar.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Error interno sin detalles expuestos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    CorrelationId:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
    UuidV4:
      type: string
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
...
    Permission:
      type: string
      enum:
        - agent:read
        - emails:manage
        - products:read
        - products:write
        - users:read
        - users:write
        - roles:read
        - roles:write
    Role:
      type: object
      additionalProperties: false
      required: [id, key, name, description, isProtected, permissions]
      properties:
        id:
          $ref: '#/components/schemas/UuidV4'
        key:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,48}$'
        name:
          type: string
          maxLength: 120
        description:
          type: string
          nullable: true
          maxLength: 500
        isProtected:
          type: boolean
        permissions:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Permission'
    User:
      type: object
      additionalProperties: false
      required: [id, email, displayName, isActive, role]
      properties:
        id:
          $ref: '#/components/schemas/UuidV4'
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          nullable: true
          maxLength: 160
        isActive:
          type: boolean
        role:
          $ref: '#/components/schemas/Role'
...
    UserEnvelope:
      type: object
      additionalProperties: false
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    UsersEnvelope:
      type: object
      additionalProperties: false
      required: [users]
      properties:
        users:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/User'
    RoleEnvelope:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role:
          $ref: '#/components/schemas/Role'
    RolesEnvelope:
      type: object
      additionalProperties: false
      required: [roles]
      properties:
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          nullable: true
          maxLength: 160
        roleId:
          $ref: '#/components/schemas/UuidV4'
    UpdateUserRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        displayName:
          type: string
          nullable: true
          maxLength: 160
        roleId:
          $ref: '#/components/schemas/UuidV4'
        isActive:
          type: boolean
    CreateRoleRequest:
      type: object
      additionalProperties: false
      required: [key, name, permissions]
      properties:
        key:
          type: string
          pattern: '^[A-Z][A-Z0-9_]{1,48}$'
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          nullable: true
          maxLength: 500
        permissions:
          type: array
          minItems: 1
          maxItems: 8
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Permission'
    UpdateRoleRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          nullable: true
          maxLength: 500
        permissions:
          type: array
          minItems: 1
          maxItems: 8
          uniqueItems: true
          items:
            $ref: '#/components/schemas/Permission'
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [correlationId, error]
      properties:
        correlationId:
          $ref: '#/components/schemas/CorrelationId'
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - BAD_REQUEST
                - CONFLICT
                - FORBIDDEN
                - INTERNAL_ERROR
                - NOT_FOUND
                - RATE_LIMITED
                - UNAUTHENTICATED
            message:
              type: string
              description: Mensaje genérico sin detalles internos.
...